Final Push
ready for review. The body is now horizontally centered, both of the navs now are their correct orientations, and take up the correct widths. The topics list now is horizontal, with the images being much smaller, and centered horizontally. The comments section will now have two columns, and a horizontal line above it, which takes up half the body. The paragraphs within the comments will be centered and floated correctly. Added comments